remote-ui/round/requirements.txt
# SecuBox Eye Remote Round - Requirements
# Pour Raspberry Pi Zero W avec HyperPixel 2.1 Round

# Touch controller (Pimoroni HyperPixel 2r)
hyperpixel2r>=0.0.1

# Fallback evdev (si kernel driver actif)
evdev>=1.6.0

# I2C access (pour debug/test)
smbus2>=0.4.0

# Image rendering
pillow>=9.0.0

# Async HTTP client
httpx>=0.23.0
aiohttp>=3.8.0

# JWT authentication
python-jose[cryptography]>=3.3.0
PyJWT>=2.6.0

# TOML config
tomli>=2.0.0

# Network utilities
netifaces>=0.11.0

# System metrics (optional, fallback to /proc)
# psutil>=5.9.0  # NOT USED - using /proc directly

# GPIO access (for interrupt pin)
RPi.GPIO>=0.7.0

# Web server (for local API)
fastapi>=0.92.0
uvicorn>=0.17.0

# WebSocket
websockets>=10.0
